+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/LandscapeProxy.cpp
@@ -181,7 +181,7 @@ void LandscapeProxy::UpdateDisplayedTexture()
 	
 	int32 tilemaskWidth = tilemaskTexture->GetWidth();
 	int32 tilemaskHeight = tilemaskTexture->GetHeight();
-	Sprite* tilemaskSprite = Sprite::CreateFromTexture(tilemaskTexture, 0, 0, tilemaskWidth, tilemaskHeight);
+	Sprite* tilemaskSprite = Sprite::CreateFromTexture(tilemaskTexture, 0, 0, (float32)tilemaskWidth, (float32)tilemaskHeight);
 	
 	Sprite *dstSprite = Sprite::CreateAsRenderTarget((float32)tilemaskWidth, (float32)tilemaskHeight, FORMAT_RGBA8888);
 	RenderManager::Instance()->SetRenderTarget(dstSprite);
@@ -199,7 +199,7 @@ void LandscapeProxy::UpdateDisplayedTexture()
 	if (notPassableTexture && texturesEnabled[TEXTURE_TYPE_NOT_PASSABLE])
 	{
 		RenderManager::Instance()->SetColor(Color::White());
-		notPassableSprite = Sprite::CreateFromTexture(notPassableTexture, 0, 0, tilemaskWidth, tilemaskHeight);
+		notPassableSprite = Sprite::CreateFromTexture(notPassableTexture, 0, 0, (float32)tilemaskWidth, (float32)tilemaskHeight);
 		notPassableSprite->SetPosition(0.f, 0.f);
 		notPassableSprite->Draw();
 	}
@@ -210,7 +210,7 @@ void LandscapeProxy::UpdateDisplayedTexture()
 	if (customColorsTexture && texturesEnabled[TEXTURE_TYPE_CUSTOM_COLORS])
 	{
 		RenderManager::Instance()->SetColor(1.f, 1.f, 1.f, .5f);
-		customColorsSprite = Sprite::CreateFromTexture(customColorsTexture, 0, 0, tilemaskWidth, tilemaskHeight);
+		customColorsSprite = Sprite::CreateFromTexture(customColorsTexture, 0, 0, (float32)tilemaskWidth, (float32)tilemaskHeight);
 		customColorsSprite->SetPosition(0.f, 0.f);
 		customColorsSprite->Draw();
 		RenderManager::Instance()->SetColor(Color::White());
@@ -222,7 +222,7 @@ void LandscapeProxy::UpdateDisplayedTexture()
 	if (visibilityCheckToolTexture && texturesEnabled[TEXTURE_TYPE_VISIBILITY_CHECK_TOOL])
 	{
 		RenderManager::Instance()->SetColor(Color::White());
-		visibilityCheckToolSprite = Sprite::CreateFromTexture(visibilityCheckToolTexture, 0, 0, tilemaskWidth, tilemaskHeight);
+		visibilityCheckToolSprite = Sprite::CreateFromTexture(visibilityCheckToolTexture, 0, 0, (float32)tilemaskWidth, (float32)tilemaskHeight);
 		visibilityCheckToolSprite->SetPosition(0.f, 0.f);
 		visibilityCheckToolSprite->Draw();
 	}
